Refactor resolve_node_ref to use normalize_module_path instead of suffix matching
- Extract normalize_module_path to its own util file for reuse
- Update resolve_node_ref to use exact matching after normalizing module path
- Move VirtualFileFinder test to ee/vellum_ee directory
